+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Close HTTP server on Boot cleanup hook
+
 ## 0.1.1
 
 - Fix: HTTP server was started each time task runs, instead of only the first time